various small fixes
