--- enforcer/nflog/interface.go
@@ -1,7 +1,0 @@
-package nflog
-
-// NFLogger is the interface of a NFLog capable struct.
-type NFLogger interface {
-	Start()
-	Stop()
-}
